+++ bodyprocessors/multipart.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The Coraza Authors
+// Copyright 2022 Juan Pablo Tosso
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -24,7 +24,6 @@ import (
 	"strings"
 
 	"github.com/corazawaf/coraza/v2/types/variables"
-	mp "github.com/corazawaf/coraza/v2/utils/multipart"
 )
 
 type multipartBodyProcessor struct {
@@ -59,7 +58,7 @@ func (mbp *multipartBodyProcessor) Read(reader io.Reader, options Options) error
 		}
 
 		// if is a file
-		filename := mp.OriginFileName(p)
+		filename := originFileName(p)
 		if filename != "" {
 			temp, err := os.CreateTemp(storagePath, "crzmp*")
 			if err != nil {
@@ -132,3 +131,18 @@ func (mbp *multipartBodyProcessor) VariableHook() variables.RuleVariable {
 var (
 	_ BodyProcessor = (*multipartBodyProcessor)(nil)
 )
+
+// OriginFileName returns the filename parameter of the Part's Content-Disposition header.
+// This function is based on (multipart.Part).parseContentDisposition,
+// See https://go.googlesource.com/go/+/refs/tags/go1.17.9/src/mime/multipart/multipart.go#87
+// for the current implementation and also notice this function hasn't change since go1.4, as in
+// https://go.googlesource.com/go/+/refs/tags/go1.4/src/mime/multipart/multipart.go#75
+func originFileName(p *multipart.Part) string {
+	v := p.Header.Get("Content-Disposition")
+	_, dispositionParams, err := mime.ParseMediaType(v)
+	if err != nil {
+		return ""
+	}
+
+	return dispositionParams["filename"]
+}
