+++ bodyprocessors/multipart_test.go
@@ -15,6 +15,8 @@
 package bodyprocessors
 
 import (
+	"mime/multipart"
+	"net/textproto"
 	"os"
 	"strings"
 	"testing"
@@ -73,5 +75,25 @@ Content-Type: text/html
 			}
 
 		}
+	}
+}
+
+func TestOriginalFileName(t *testing.T) {
+	tests := map[string][2]string{
+		"no filename":       {` form-data ; name=foo`, ""},
+		"contains filename": {`form-data; name="file"; filename="test.txt"`, "test.txt"},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			p := &multipart.Part{
+				Header: textproto.MIMEHeader{
+					"Content-Disposition": []string{test[0]},
+				},
+			}
+			if got, want := originFileName(p), test[1]; got != want {
+				t.Errorf("OriginFileName(%v) = %v, want %v", p, got, want)
+			}
+		})
 	}
 }
