+++ seclang/rules_test.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The Coraza Authors
+// Copyright 2022 Juan Pablo Tosso
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
