--- utils/multipart/multipart.go
@@ -1,35 +1,0 @@
-// Copyright 2022 The Coraza Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package multipart
-
-import (
-	"mime"
-	"mime/multipart"
-)
-
-// OriginFileName returns the filename parameter of the Part's Content-Disposition header.
-// This function is based on (multipart.Part).parseContentDisposition,
-// See https://go.googlesource.com/go/+/refs/tags/go1.17.9/src/mime/multipart/multipart.go#87
-// for the current implementation and also notice this function hasn't change since go1.4, as in
-// https://go.googlesource.com/go/+/refs/tags/go1.4/src/mime/multipart/multipart.go#75
-func OriginFileName(p *multipart.Part) string {
-	v := p.Header.Get("Content-Disposition")
-	_, dispositionParams, err := mime.ParseMediaType(v)
-	if err != nil {
-		return ""
-	}
-
-	return dispositionParams["filename"]
-}
